backend/requirements.txt
# Core FastAPI dependencies
fastapi
uvicorn[standard] # asgi server with extra dependencies
pydantic # data validation
python-dotenv # environment variables

# Basic HTTP and data handling
requests #http reuqests
httpx # http client

# Web scraping and parsing
beautifulsoup4 # web scraping
lxml  # XML/HTML parser for BeautifulSoup

# AI/LLM
anthropic # claude
google-generativeai # gemini

# Development and testing
python-multipart  # handle file uploads + form data
pytest  # testing framework
pytest-asyncio # async testing support

# Additional utilities
typing-extensions  # enhanced typing support
